fix status for invalid private keys
