+++ src/core/models.py
@@ -184,6 +184,11 @@ class EtcdServer(RelationState):
         """Get the current progress of the restore workflow."""
         return RestoreStep(self.relation_data.get("restore_step", ""))
 
+    @property
+    def is_leader(self) -> bool:
+        """Check if the current unit is the leader of the cluster."""
+        return self.unit.is_leader()
+
 
 class EtcdCluster(RelationState):
     """State/Relation data collection for the etcd application."""
